Dynamic round boudning, do soemthing different for DQNs made prior to normalization, etc.

+++ web/auctions/auctions/management/commands/nfsp.py
@@ -1,6 +1,7 @@
 from django.core.management.base import BaseCommand
 from open_spiel.python.examples.ubc_nfsp_example import run_nfsp, setup_directory_structure
 from open_spiel.python.examples.ubc_utils import smart_load_sequential_game, load_game_config, fix_seeds, read_config, apply_optional_overrides, add_optional_overrides, default_device
+from open_spiel.python.examples.ubc_decorators import UniformRestrictedNashResponseAgent, UniformRestrictedNashResponseAgentSelector
 import sys
 import logging
 import pickle
@@ -34,6 +35,14 @@ class DBNFSPSaver:
         
         return episode
 
+class NullResultSaver:
+
+    def __init__(self) -> None:
+        pass
+    
+    def save(self, result):
+        pass
+
 class DBBRDispatcher:
 
     def __init__(self, num_players, eval_overrides, br_overrides, eq_solver_run, br_portfolio_path):
@@ -50,6 +59,14 @@ class DBBRDispatcher:
             dispatch.dispatch_eval_database(eq.experiment.name, eq.name, t, player, 'straightforward', overrides=self.eval_overrides) # Straightforward eval
         dispatch.dispatch_eval_database(eq.experiment.name, eq.name, t, None, None, overrides=self.eval_overrides)
             
+class NullDispatcher:
+
+    def __init__(self) -> None:
+        pass
+
+    def dispatch(self, t):
+        pass
+
 
 class Command(BaseCommand):
     help = 'Runs NFSP and saves the results'
@@ -63,6 +80,7 @@ class Command(BaseCommand):
         parser.add_argument('--device', type=str, default=default_device)
 
         parser.add_argument('--overwrite_db', type=util.strtobool, default=0)
+        parser.add_argument('--dry_run', type=util.strtobool, default=0)
 
         # Directory
         parser.add_argument('--output_dir', type=str, default='output') # Note: DONT NAME THIS "checkpoints" because of a jupyter notebook
@@ -73,7 +91,7 @@ class Command(BaseCommand):
         parser.add_argument('--job_name', type=str, default='auction')
         parser.add_argument('--filename', type=str, default='parameters.json') # Select clock auction game
         parser.add_argument('--game_name', type=str, default='clock_auction')
-        
+
         # Reporting and evaluation
         parser.add_argument('--report_freq', type=int, default=50_000)
         parser.add_argument('--eval_every', type=int, default=300_000)
@@ -86,6 +104,11 @@ class Command(BaseCommand):
         parser.add_argument('--br_portfolio_path', type=str, default=None)
         parser.add_argument('--br_overrides', type=str, default='')
         parser.add_argument('--eval_overrides', type=str, default='')
+
+        # RNR
+        parser.add_argument('--rnr_player', type=int, default=None)
+        parser.add_argument('--rnr_exploit_prob', type=float, default=0.5)
+        parser.add_argument('--rnr_checkpoints', type=int, nargs="+", default=[])
 
         add_optional_overrides(parser)
 
@@ -130,30 +153,46 @@ class Command(BaseCommand):
                 config=game_config
             )
 
-        # 2) Make the experiment if it doesn't exist
-        experiment, _ = Experiment.objects.get_or_create(name=experiment_name)
+        if not opts.dry_run:
+            # 2) Make the experiment if it doesn't exist
+            experiment, _ = Experiment.objects.get_or_create(name=experiment_name)
 
-        output_dir = f'{OUTPUT_ROOT}/{experiment_name}/{run_name}'
-        setup_directory_structure(output_dir, opts.warn_on_overwrite)
+            output_dir = f'{OUTPUT_ROOT}/{experiment_name}/{run_name}'
+            setup_directory_structure(output_dir, opts.warn_on_overwrite)
 
-        # Save the game config so there's no confusion later if you need to cross-reference. Shouldn't techincally need this in the database version, but why not
-        with open(f'{output_dir}/game.json', 'w') as outfile:
-            json.dump(game.config, outfile)
+            # Save the game config so there's no confusion later if you need to cross-reference. Shouldn't techincally need this in the database version, but why not
+            with open(f'{output_dir}/game.json', 'w') as outfile:
+                json.dump(game.config, outfile)
 
+            if overwrite_db:
+                try:
+                    EquilibriumSolverRun.objects.get(experiment=experiment, name=run_name, game=game).delete()
+                except EquilibriumSolverRun.DoesNotExist:
+                    pass
 
-        if overwrite_db:
-            try:
-                EquilibriumSolverRun.objects.get(experiment=experiment, name=run_name, game=game).delete()
-            except EquilibriumSolverRun.DoesNotExist:
-                pass
+            # 3) Make an EquilibriumSolverRun
+            eq_solver_run = EquilibriumSolverRun.objects.create(experiment=experiment, name=run_name, game=game, config=config)
 
-        # 3) Make an EquilibriumSolverRun
-        eq_solver_run = EquilibriumSolverRun.objects.create(experiment=experiment, name=run_name, game=game, config=config)
+            # Load the environment from the database
+            env_and_model = env_and_model_from_run(eq_solver_run)
+        else:
+            env_and_model = env_and_model_for_dry_run(game, config)
 
-        # Load the environment from the database
-        env_and_model = env_and_model_from_run(eq_solver_run)
+        if opts.rnr_player is not None:
+            if len(opts.rnr_checkpoints) == 0:
+                raise ValueError("Must supply rnr checkpoints")
+            
+            db_checkpoints = EquilibriumSolverRunCheckpoint.objects.filter(pk__in=opts.rnr_checkpoints)
+            other_env_and_models = [db_checkpoint_loader(db_checkpoint) for db_checkpoint in db_checkpoints]
+            agent_selector = UniformRestrictedNashResponseAgentSelector(len(db_checkpoints), game.num_players, exploit_prob=opts.rnr_exploit_prob, iterate_br=opts.iterate_br, rnr_player_id=opts.rnr_player) 
+            for player in range(game.num_players):
+                if player == opts.rnr_player:
+                    continue
+                env_and_model.agents[player] = UniformRestrictedNashResponseAgent(env_and_model.agents[player], [env.agents[player] for env in other_env_and_models], agent_selector)
+        else:
+            agent_selector = None
 
-        result_saver = DBNFSPSaver(eq_solver_run=eq_solver_run)
-        dispatcher = DBBRDispatcher(game.num_players, opts.eval_overrides, opts.br_overrides, eq_solver_run, opts.br_portfolio_path)
+        result_saver = DBNFSPSaver(eq_solver_run=eq_solver_run) if not opts.dry_run else NullResultSaver()
+        dispatcher = DBBRDispatcher(game.num_players, opts.eval_overrides, opts.br_overrides, eq_solver_run, opts.br_portfolio_path) if not opts.dry_run else NullDispatcher()
 
-        run_nfsp(env_and_model, opts.num_training_episodes, opts.iterate_br, result_saver, seed, opts.compute_nash_conv, dispatcher, opts.eval_every, opts.eval_every_early, opts.eval_exactly, opts.eval_zero, opts.report_freq, opts.dispatch_br)
+        run_nfsp(env_and_model, opts.num_training_episodes, opts.iterate_br, result_saver, seed, opts.compute_nash_conv, dispatcher, opts.eval_every, opts.eval_every_early, opts.eval_exactly, opts.eval_zero, opts.report_freq, opts.dispatch_br, agent_selector)
